Attempt to get repo to build on another machine
This is not working, but it fails with a message which suggests
that the input geometries are incorrect.

I will try to detect this and try again, but something is still wrong.

+++ extract-shapefiles.py
@@ -3,11 +3,14 @@ import zipfile
 fp = 'preprocessing/input/osm/'
 s = os.listdir(fp)
 for n in s:
+    if not '.zip' in n:
+        continue
     p = n.split('-')
     print(n)
     print(p[0])
     z = zipfile.ZipFile(f"{fp}/{n}")
     z.extractall(path=fp)
     for zmn in z.namelist():
-        os.rename(f"{fp}/{zmn}",
-                f"{fp}/{p[0]}_{zmn}")
+        if not os.path.exists(f"{fp}/{p[0]}_{zmn}"):
+            os.rename(f"{fp}/{zmn}",
+                    f"{fp}/{p[0]}_{zmn}")
